添加用户管理等

+++ src/flight.rs
@@ -1,15 +1,15 @@
 use crate::initdb::DbInfo;
 pub struct Flight {
-    num: u32,
-    capacity: u32,
-    booked: u32,
+    num: i32,
+    capacity: i32,
+    booked: i32,
     price: f64,
     leave_city: Option<String>,
     arrive_city: Option<String>,
     leave_airport: Option<String>,
     arrive_airport: Option<String>,
-    leave_time: u32,
-    arrive_time: u32,
+    leave_time: i64,
+    arrive_time: i64,
 }
 
 impl Flight {
